django_app/requirements.txt
Django>=4.0
psycopg2-binary>=2.9
pandas>=1.3.0
openpyxl>=3.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0